+++ docs/configuration.md
@@ -109,7 +109,7 @@ Unless you need to set a non-default value, it is recommended to only populate o
 | `BATCH_REQUESTS_DISALLOWED_METHODS` | ["debug_traceTransaction", "eth_newFilter", "eth_uninstallFilter", "eth_getFilterChanges", "eth_getFilterLogs", "eth_newBlockFilter", "eth_newPendingTransactionFilter"] | A list of the methods that are not allowed for batch.                                                                                                                                                                             |
 | `BATCH_REQUESTS_ENABLED`            | "true"                                                                                                                                                                   | Flag to disable or enable batch requests.                                                                                                                                                                                         |
 | `BATCH_REQUESTS_MAX_SIZE`           | "100"                                                                                                                                                                    | Maximum number of requests allowed in a batch.                                                                                                                                                                                    |
-| `DISABLED_RPC_METHODS`              | "[]"                                                                                                                                                                     | A list of RPC methods that can be disabled by relay's operator.                                                                                                                                                                   |
+| `DISABLE_HEDERA_NAMESPACE`          | "false"                                                                                                                                                                  | The JSON-RPC Relay provide custom methods under the `hedera_` namespace. This flag enables or disables access to this namespace. Enabled by default.                                                                              |
 | `CHAIN_ID`                          | ""                                                                                                                                                                       | The network chain id. Local and previewnet envs should use `0x12a` (298). Previewnet, Testnet and Mainnet should use `0x129` (297), `0x128` (296) and `0x127` (295) respectively.                                                 |
 | `HEDERA_NETWORK`                    | ""                                                                                                                                                                       | Which network to connect to. Automatically populates the main node & mirror node endpoints. Can be `previewnet`, `testnet`, `mainnet` or a map of network IPs -> node accountIds e.g. `{"127.0.0.1:50211":"0.0.3"}`               |
 | `INPUT_SIZE_LIMIT`                  | "1mb"                                                                                                                                                                    | The [koa-jsonrpc](https://github.com/Bitclimb/koa-jsonrpc) maximum size allowed for requests                                                                                                                                      |
